코드 분할과 빌드 sh 추가
