+++ .gitignore
@@ -1,0 +1,80 @@
+# Build artifacts
+*.d                       # Dependency files generated by build tools
+*.o                       # Object files
+*.so                      # Shared object files
+*.a                       # Static library files
+*.dll                     # Windows dynamic link libraries
+*.exe                     # Executable files (Windows)
+/*.exe                    # Executable in the root directory
+*.out                     # Linux/macOS executables
+*.app                     # macOS application bundles
+
+# Debugging and profiling
+*.pdb                     # Windows debug symbols
+*.profraw                 # Profiling data files
+*.profdata                # Profiling result files
+*.dSYM/                   # macOS debug symbols
+*.log                     # Logs (general debugging or build logs)
+*.trace                   # Debug trace files
+
+# Operating System specific files
+.DS_Store                 # macOS Finder metadata
+Thumbs.db                 # Windows thumbnail cache
+desktop.ini               # Windows desktop configuration
+*.swp                     # Swap files (Vim, etc.)
+*.swo                     # Vim temporary files
+~$*                       # Temporary files created by some text editors
+
+# IDEs and Editors
+.idea/                    # IntelliJ IDEA
+*.iml                     # IntelliJ module files
+.vscode/                  # Visual Studio Code
+*.code-workspace          # VS Code workspace files
+*.sublime-project         # Sublime Text project files
+*.sublime-workspace       # Sublime Text workspace files
+
+# Temporary Files
+*.tmp                     # Temporary files
+*.bak                     # Backup files
+*.old                     # Backup files (alternative)
+*.orig                    # Merge conflict backups
+*.rej                     # Patch rejects
+/tmp/                     # Temporary directories
+/temp/                    # Alternative temp directories
+
+# Archives
+*.tar
+*.gz
+*.zip
+*.rar
+*.7z
+*.bz2
+
+# Testing and Coverage
+*.test                    # Test artifacts
+*.snap                    # Snapshot files
+coverage/                 # Coverage reports
+*.lcov                    # LCov coverage files
+
+# Dependency directories
+node_modules/             # Node.js dependencies (if applicable)
+vendor/                   # Vendor dependencies (e.g., Go, PHP)
+
+# Environment and sensitive files
+*.env                     # Environment variables
+.env.*                    # Other .env variants
+*.key                     # Private keys
+*.pem                     # Certificates
+*.crt                     # Certificates
+*.pfx                     # PKCS12 certificate files
+
+# System specific caches
+.cache/                   # System or application cache
+.lock/                    # Lock files
+
+# Build tools and configs
+Makefile                  # If autogenerated
+CMakeFiles/               # CMake build artifacts
+build/                    # General build output folder
+cmake-build-debug/        # Debug builds from CMake
+cmake-build-release/      # Release builds from CMake
