+++ linker.ld
@@ -1,0 +1,24 @@
+ENTRY(_start)
+
+SECTIONS
+{
+    . = 0x8000;
+
+    .text : {
+        *(.text.start)
+        *(.text)
+    }
+
+    .rodata : {
+        *(.rodata)
+    }
+
+    .data : {
+        *(.data)
+    }
+
+    .bss : {
+        *(COMMON)
+        *(.bss)
+    }
+}
