+++ src/kernel/command.h
@@ -1,0 +1,9 @@
+#ifndef COMMAND_H
+#define COMMAND_H
+
+#include "dc.h"
+
+void process_command(const char *cmd);
+
+
+#endif //COMMAND_H
