+++ src/kernel/system.h
@@ -1,0 +1,8 @@
+#ifndef SYSTEM_H
+#define SYSTEM_H
+
+void sysinfo();
+void reboot_system();
+void shutdown_system();
+
+#endif //SYSTEM_H
